Revert non-atomic Node count read+write when Node is full
The previous commit attempted to remove the getAndIncrement when a Node
is full. However, this introduces a subtle bug when the Node is full.
Imagine two threads are adding items to a full Node. The first caller
reads the volatile count, and sees it is greater than or equal to the
Node capacity. It then constructs the next tail candidate, and then
CASes the Node's next value. However, if the second thread wins the CAS
(e.g. because the construction of its next Node candidate took less time)
the second thread will add its value prior to the first caller. The
first thread will re-read the tail and observed its been updated (by
thread two) and append its item after the second thread's value in the
new tail Node (generally, more retries may apply).